+++ changelog.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [unreleased]
 
+- Adjuster respects diagram style.Spacing 
 - NewDiagram returns a pointer to diagram
 - Diagram variations embed reference to diagram